Ajout package cookie et auth avec version dev récente
