+++ src/main/java/com/wildcodeschool/webook/Auth/domain/dto/UserPrincipal.java
@@ -19,7 +19,7 @@ public class UserPrincipal implements UserDetails {
     public Collection<? extends GrantedAuthority> getAuthorities(){
         List<GrantedAuthority> authorities = new ArrayList<>();
         authorities.add(new SimpleGrantedAuthority(user.getRole().getType()));
-                return authorities;
+        return authorities;
     }
 
     @Override
