+++ src/main/java/com/wildcodeschool/webook/Auth/domain/entity/User.java
@@ -127,4 +127,6 @@ public class User {
     public void setBooks(List<Book> books) {
         this.books = books;
     }
+
+
 }
