+++ src/main/java/com/wildcodeschool/webook/Auth/infrastructure/repository/UserRepository.java
@@ -5,4 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     User findByEmail(String email);
+    User findByUsername(String username);
 }
