--- src/main/java/com/wildcodeschool/webook/auth/application/ApiConfig.java
@@ -1,17 +1,0 @@
-package com.wildcodeschool.webook.auth.application;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class ApiConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**") // Tous les chemins
-                .allowedOrigins("http://localhost:4200") // URL(s) autorisée(s)
-                .allowedMethods("GET", "POST", "PUT", "DELETE") // Méthodes HTTP autorisées
-                .allowCredentials(true)
-                .maxAge(3600);
-    }
-}
