--- src/main/java/com/wildcodeschool/webook/auth/domain/dto/UserDTO.java
@@ -1,85 +1,0 @@
-package com.wildcodeschool.webook.auth.domain.dto;
-
-import com.wildcodeschool.webook.auth.domain.entity.Role;
-
-public class UserDTO {
-    private Long id;
-
-    private String firstName;
-
-    private String lastName;
-
-    private String username;
-
-    private String email;
-
-    private String password;
-
-    private Boolean isEnabled;
-
-    private Role roles;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public Boolean getEnabled() {
-        return isEnabled;
-    }
-
-    public void setEnabled(Boolean enabled) {
-        isEnabled = enabled;
-    }
-
-    public Role getRoles() {
-        return roles;
-    }
-
-    public void setRoles(Role roles) {
-        this.roles = roles;
-    }
-}
