--- src/main/java/com/wildcodeschool/webook/auth/domain/entity/Role.java
@@ -1,38 +1,0 @@
-package com.wildcodeschool.webook.auth.domain.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String type;
-
-    public Role() {
-    }
-
-    public Role(String type) {
-        this.type = type; // ADMIN, USER, ...
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-}
