--- src/main/java/com/wildcodeschool/webook/auth/infrastructure/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.wildcodeschool.webook.auth.infrastructure.repository;
-
-import com.crudblog.demo.auth.domain.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUsername(String username);
-}
